+++ src/test/java/visiblehand/parser/air/AAParserTest.java
@@ -30,75 +30,75 @@ public class AAParserTest extends EbeanTest {
 	public void testGetAirport() throws ParseException {
 		AAParser parser = new AAParser();
 		Airport airport = parser.getAirport("RALEIGH DURHAM");
-		assertEquals(airport.getId(), 3626);
+		assertEquals(airport.getId().longValue(), 3626);
 		
 		airport = parser.getAirport("ALBUQUERQUE");
-		assertEquals(airport.getId(), 4019);
+		assertEquals(airport.getId().longValue(), 4019);
 		
 		airport = parser.getAirport("LOS ANGELES");
-		assertEquals(airport.getId(), 3484);
+		assertEquals(airport.getId().longValue(), 3484);
 		
 		airport = parser.getAirport("WASHINGTON REAGAN");
-		assertEquals(airport.getId(), 3520);
+		assertEquals(airport.getId().longValue(), 3520);
 		
 		airport = parser.getAirport("DALLAS FT WORTH");
-		assertEquals(airport.getId(), 3670);
+		assertEquals(airport.getId().longValue(), 3670);
 		
 		airport = parser.getAirport("AUSTIN");
-		assertEquals(airport.getId(), 3673);
+		assertEquals(airport.getId().longValue(), 3673);
 		
 		airport = parser.getAirport("NEW YORK JFK");
-		assertEquals(airport.getId(), 3797);
+		assertEquals(airport.getId().longValue(), 3797);
 		
 		airport = parser.getAirport("NEW YORK LGA");
-		assertEquals(airport.getId(), 3697);
+		assertEquals(airport.getId().longValue(), 3697);
 		
 		airport = parser.getAirport("CHICAGO OHARE");
-		assertEquals(airport.getId(), 3830);
+		assertEquals(airport.getId().longValue(), 3830);
 		
 		airport = parser.getAirport("NEWARK");
-		assertEquals(airport.getId(), 3494);
+		assertEquals(airport.getId().longValue(), 3494);
 		
 		airport = parser.getAirport("SALT LAKE CITY");
-		assertEquals(airport.getId(), 3536);
+		assertEquals(airport.getId().longValue(), 3536);
 		
 		airport = parser.getAirport("LONDON HEATHROW");
-		assertEquals(airport.getId(), 507);
+		assertEquals(airport.getId().longValue(), 507);
 		
 		airport = parser.getAirport("BARCELONA");
-		assertEquals(airport.getId(), 1218);
+		assertEquals(airport.getId().longValue(), 1218);
 		
 		airport = parser.getAirport("SAN JUAN       PR");
-		assertEquals(airport.getId(), 2890);
+		assertEquals(airport.getId().longValue(), 2890);
 		
 		airport = parser.getAirport("SARASOTA/BRADENTN");
-		assertEquals(airport.getId(), 4067);
+		assertEquals(airport.getId().longValue(), 4067);
 		
 		airport = parser.getAirport("HONOLULU");
-		assertEquals(airport.getId(), 3728);
+		assertEquals(airport.getId().longValue(), 3728);
 		
 		airport = parser.getAirport("LAS VEGAS");
-		assertEquals(airport.getId(), 3877);
+		assertEquals(airport.getId().longValue(), 3877);
 		
 		airport = parser.getAirport("SAN FRANCISCO");
-		assertEquals(airport.getId(), 3469);
+		assertEquals(airport.getId().longValue(), 3469);
 		
 		airport = parser.getAirport("ORLANDO INTL");
-		assertEquals(airport.getId(), 3878);
+		assertEquals(airport.getId().longValue(), 3878);
 		
 		airport = parser.getAirport("CHAMPAIGN");
-		assertEquals(airport.getId(), 4049);
+		assertEquals(airport.getId().longValue(), 4049);
 		
 		airport = parser.getAirport("BOSTON");
-		assertEquals(airport.getId(), 3448);
+		assertEquals(airport.getId().longValue(), 3448);
 		
 		airport = parser.getAirport("MIAMI INTERNTNL");
-		assertEquals(airport.getId(), 3576);
+		assertEquals(airport.getId().longValue(), 3576);
 		
 		airport = parser.getAirport("SANTA ANA");
-		assertEquals(airport.getId(), 3867);
+		assertEquals(airport.getId().longValue(), 3867);
 		
 		airport = parser.getAirport("MONTREAL TRUDEAU");
-		assertEquals(airport.getId(), 146);
+		assertEquals(airport.getId().longValue(), 146);
 	}
 }
